Fix C90 compliance and long int support
- Remove C++ style comments (use only /* */)
- Add support for long integers in handle_d_i
- Increase buffer size for large numbers
- Maintain ISO C90 compatibility

+++ main.h
@@ -5,6 +5,6 @@
 
 int _printf(const char *format, ...);
 int _putchar(char c);
-int handle_d_i(int num);
+int handle_d_i(long num);
 
 #endif
